Allow running of just a subset of the SQL statement tests.
It takes a long time to run all the tests - this allows one
or more tests to be specified on the command line. Only those
tests will be run. If no test names are specified, all the 
tests are run (as current behaviour).

FossilOrigin-Name: 91ec6ff6415fe0480ec7b84523e9a6eabacd9824