Build with conan dependencies and locating code
This commit adds some changes to fix the build in case of static linkage
(a flag was missing, I had to write a fake geos-config script) both for
windows and *unix.

I also had to add some code to "locate" easily (at runtime) where the
dll is, so that I can use to find the real path and load it to sqlite. I
think this last part can be eventually removed, but for now it is easier
this way.