Using CMake instead of configure
For a library that aims to work on windows and on linux and on macos,
using a cross-platform build tool makes things easier. This made it
possible to statically compile the dependencies for spatialite, using a
package manager (conan) to explicitly define versions of them. It should
still be possible to compile against the system-wide dependencies by
tweaking the CMakeLists.txt, but I don't think it's worthy for our
use-case for now (and by the way it's what configure/nmake scripts are
doing right now, except that on Windows you need explicit paths for
dependecies, which is as ugly as it can get). The usage of conan.io made
it possible to neable RTTOPO on windows as well.

We also added a location.cpp file to be used with C# (our target) to
find the real path of the library, given that it's not possible to do
any better untile we have support from System.Data.SQLite guys - which
seems unlikely.