Adding PROJ as dependency
We added PROJ as a submodule dependency as it is clear that using conan
is not the right way to do this. Of course this is true just because
proj is not meant to be an extension or a standalone library and links
directly to libsqlite3. It is funny to see that most of the usage of
sqlite is `sqlite3_vsnprintf` (!). Great perfromance gains here! :-).
This dependency, once included in spatialite creates a dependency loop.
Despite being an accepted bad-practice, on a normal linux installation
"works" if no dlopen is involved (except for sqlite3_load_extension), as
linux is very lax (and thus powerful) when it comes to symbol
resolution: you would normally load libsqlite3 than dlopen spatialite
which causes libproj to be loaded as well. Than sice libsqlite3 uses
the RTLD_GLOBAL, it will be possible for libproj to link against the
already-loaded lib and everything will be fine.

Turns out this would normally work on windows, too, if using a "normal"
libsqlite3 dll (for similar reasons). It is thoug a bit perverted to use
sqlite3 this way: functions are defined (and called) differently
throughout the application depending if a call comes from spatialite or
proj:

 - proj uses directly linked (by the loader) calls to sqlite3 functions;
 - spatialite initializes a struct with function-pointers to the very
   same functions and calls them with an indirect call - which is the
   reccomended way to use sqlite3 in an extension.

As a side not, this method, which is slightly less performant was designed
by sqlite3 with the very intent of breaking dependency loops: spatialite,
linking with proj made this effor useless.

Coming to C#... System.Data.SQLite uses an amalgamation of sqlite with
some additions to support P/Invoke calls. The compiled this library with
the name "SQLite.Interop.dll" (the extension is .dll on macos and linux,
too, probably to support older framework versions). This is a problem as
the loop is now broken. Now SQLite.Interop.dll loads mod_spatialite
which loads libsqlite3 (again). You either find libsqlite3 and load two
copies of sqlite in memory (which is
[BAD](https://www.sqlite.org/howtocorrupt.html)), otherwise it will just
fail to load.

I don't phantom to understand why this naming decision was taken, but
that's what we have. So we had to change PROJ to behave like an
extension, too. We had also to remove some "optimizations" as they were
called during library startup (c++ static initializers which allocates
and call sqlite3_vsnprintf... yuk!).

We also had to remove all the usages of the rtree functionality (just
two functions and the knn module) which luckily we don't use. It would
be nice for sqlite3 to make these functions available to extensions as
well (just like everything else), but the functionality is 5 years
old... If they didn't in 5 years I doubt it will come quickly.